Implemented *Memory command.
The *Memory command is now implemented as another of the accessor
focused functions. The *DisFile tool can now be used for dumping
as well. *Dump could be created in the same way.

It looks like the dump is cut off at the end of the output for
some reason. Needs investigating.