Completed the breakpoint prefetch handler.
The prefetch handler now executes correctly and can be used to breakpoint
lines and continue afterwards. The R13/R14 values in the block have now
been set up properly and are correct for the mode we returned to. The
instructions are executed out of line. This may cause problems for PC
relative instructions. I think that's similar to how RISC OS Classic
handled things.