Initial 64bit work; removed redundant assembler; build example.
This is the initial 64bit version of the disassembler. The library
builds and the example program works. However, the DisFile tool is
showing up problems with the disassembly of some ROR instructions.
These can be fixed up in later changes.

The old assembler code that embedded the CPU register definition
for 32bit is now in C. The stdint include locally is no longer needed
as even the 32bit environment has stdint.