Fix for 64-bit architecture register header (*ShowRegs fix).
The architecture register header has a definition of the register
dump table for the current architecture. However, this was not
using the correct types, so the 64-bit reads were getting the wrong
values. *ShowRegs would then crash. It's now working ok and we
can see the registers and state properly!