Update the ARMv7 tables ordering to ensure that shifts are handled.
The shifts were not handled properly, because we never revert back
to the start of the line and re-parse the argument table. Because
of this, we need to ensure that the order of the instructions is
correct. The tables have correct orders for the LSL, LSR, ASR, ROR
shift instructions so that they're handled properly. The main reason
for the differences was the use of 'set' to remove duplicates from
a list in Python.